Refactor main kinect feeds. Option for all bodies and tracked bodies.
