+++ index.html
@@ -5,7 +5,7 @@
     <title>Kinectron</title>
 
     <script src="js/peer.min.js"></script>
-    <script src='p5examples.js' type="text/JavaScript"></script>
+    <script src="kinectron.js" type="text/JavaScript"></script>
     <link rel="stylesheet" type="text/css" href="css/style.css">
   </head>
 
@@ -45,7 +45,7 @@
           <input type="button" class="feed-option" id="le-infrared" value="Long Exposure Infrared">
           <input type="button" class="feed-option" id="key" value="Key">
           <input type="button" class="feed-option" id="fh-joint" value="Floor Height Joint">
-          <input type="button" class="feed-option" id="scale" value="Scale User">
+          <!-- <input type="button" class="feed-option" id="scale" value="Scale User"> -->
           <input type="button" class="feed-option" id="stop-all" value="Stop All">
         </div>  
         <div style="clear:both"></div>
