+++ kinectron.js
@@ -1,14 +1,23 @@
 var Kinect2 = require('kinect2');
 var kinect = new Kinect2();
 
-var mypeerid = null;
-var peer = null;
+//  Create local peer server
+var PeerServer = require('peer').PeerServer;
+var server = PeerServer({port: 9001, path: '/'});
+
+// Set peer credentials for localhost by default
+var peerNet = {host: 'localhost', port: 9001, path: '/'};
+var myPeerId = 'kinectron';
 var peer_ids = [];
 var peer_connections = [];
+var peer = null;
 var peerIdDisplay = null;
+var newPeerEntry = false;
+var newPeerInfo;
 
 var canvas = null;
 var context = null;
+var canvasState = null;
 
 var outputCanvas = null;
 var outputContext = null;
@@ -19,12 +28,20 @@ var COLORHEIGHT = 1080;
 var DEPTHWIDTH = 512;
 var DEPTHHEIGHT = 424;
 
+var outputColorW = 960;
+var outputColorH = 540;
+
+var outputDepthW = 512;
+var outputDepthH = 424;
+
 var imageData = null;
 var imageDataSize = null;
 var imageDataArray = null;
 
 var busy = false;
 var currentCamera = null;
+
+var sendAllBodies = false;
 
 // Key Tracking needs cleanup
 var trackedBodyIndex = -1;
@@ -39,22 +56,26 @@ var HANDLASSOCOLOR = 'blue';
 window.addEventListener('load', initpeer);
 window.addEventListener('load', init);
 
-function chooseCamera(camera) {
+function chooseCamera(evt, feed) {
+  var camera;
+
+  if (evt) {
+    camera = evt.srcElement.id;
+  } else {
+    camera = feed;
+  }
+
+  console.log('got one', camera);
+
   if (currentCamera) {
-    console.log('stopping');
-    console.log(currentCamera);
     changeCameraState(currentCamera, 'stop');
   }
 
   if (currentCamera == camera || camera == 'stop-all') {
-    console.log('resetting');
-    console.log(currentCamera);
     currentCamera = null;
-    return
+    return;
   }
 
-  console.log('starting');
-  console.log(currentCamera);
   changeCameraState(camera, 'start');
   currentCamera = camera;
 }
@@ -93,7 +114,13 @@ function changeCameraState(camera, state) {
       cameraCode = 'ScaleUser';
     break;
 
+    case 'body':
+      sendAllBodies = true;
+      cameraCode = 'SkeletonTracking';
+    break;
+
     case 'skeleton':
+      sendAllBodies = false;
       cameraCode = 'SkeletonTracking';
     break;
   }
@@ -115,30 +142,36 @@ function init() {
 
   peerIdDisplay = document.getElementById('peerid');
 
+  document.getElementById('peersubmit').addEventListener('click', newPeerServer);
   document.getElementById('loadfile').addEventListener('change', loadFile);
-  document.getElementById('rgb').addEventListener('click', function() {chooseCamera('rgb')});
-  document.getElementById('depth').addEventListener('click', function() {chooseCamera('depth')});
-  document.getElementById('key').addEventListener('click', function() {chooseCamera('key')});
-  document.getElementById('infrared').addEventListener('click', function() {chooseCamera('infrared')});
-  document.getElementById('le-infrared').addEventListener('click', function() {chooseCamera('le-infrared')});
-  document.getElementById('fh-joint').addEventListener('click', function() {chooseCamera('fh-joint')});
-  document.getElementById('scale').addEventListener('click', function() {chooseCamera('scale')});
-  document.getElementById('skeleton').addEventListener('click', function() {chooseCamera('skeleton')});
-  document.getElementById('stop-all').addEventListener('click', function() {
-    chooseCamera('stop-all')});
+  document.getElementById('colorwidth').addEventListener('change', updateDimFields);
+  document.getElementById('colorheight').addEventListener('change', updateDimFields);
+  document.getElementById('depthwidth').addEventListener('change', updateDimFields);
+  document.getElementById('depthheight').addEventListener('change', updateDimFields);
+  document.getElementById('colorsubmit').addEventListener('click', setOutputDimensions);
+  document.getElementById('depthsubmit').addEventListener('click', setOutputDimensions);
+  document.getElementById('rgb').addEventListener('click', chooseCamera);
+  document.getElementById('depth').addEventListener('click', chooseCamera);
+  document.getElementById('key').addEventListener('click', chooseCamera);
+  document.getElementById('infrared').addEventListener('click', chooseCamera);
+  document.getElementById('le-infrared').addEventListener('click', chooseCamera);
+  document.getElementById('fh-joint').addEventListener('click', chooseCamera);
+  //document.getElementById('scale').addEventListener('click', chooseCamera);
+  document.getElementById('skeleton').addEventListener('click', chooseCamera);
+  document.getElementById('stop-all').addEventListener('click', chooseCamera);
 }
 
 function initpeer() {
-    peer = new Peer('l',{host: 'liveweb.itp.io', port: 9000, path: '/', secure: true});
-
+    peer = new Peer(myPeerId, peerNet);
     peer.on('error',function(err) {
       console.log(err);
     });
 
     peer.on('open', function(id) {
-      console.log('My peer ID is: ' + id);
-      mypeerid = id;
-      peerIdDisplay.innerHTML = mypeerid;
+      myPeerId = id;
+      peerIdDisplay.innerHTML = myPeerId;
+      document.getElementById('port').innerHTML = peer.options.port;
+      document.getElementById('connectionopen').style.display = 'block';
   });
 
   peer.on('connection', function(conn) {
@@ -148,18 +181,54 @@ function initpeer() {
 
     connection.on('open', function() {
       console.log("Connection opened.");
+      sendToPeer('ready', {});
     });
     connection.on('data', function(data) {
       console.log("Data Received: " + data);
     });
 
     connection.on('data', function(dataReceived) {
-      if (dataReceived.event == 'feed') {
-        console.log(dataReceived.data);
-        chooseCamera(dataReceived.data);
+      if (dataReceived.event == 'initfeed') {
+        console.log(dataReceived.data.feed);
+
+        if (dataReceived.data.feed) {
+          console.log('yes got feed');
+          chooseCamera(null, dataReceived.data.feed);
+        } else {
+          console.log('no feed not setting one');
+        }
+      } else if (dataReceived.event == 'feed') {
+        chooseCamera(null, dataReceived.data.feed);
       }
     });
+
   });
+
+  peer.on('close', function() {
+    console.log('closed');
+
+    // Only create new peer if old peer destroyed and new peer requested
+    if (newPeerEntry) {
+      peer = null;
+      initpeer();
+      newPeerEntry = false;
+    }
+  });
+}
+
+
+function newPeerServer(evt) {
+  console.log('newpeerserver');
+  newPeerEntry = true;
+  evt.preventDefault();
+  myPeerId = document.getElementById('newpeerid').value;
+  var peerNetTemp = document.getElementById('peernet').value;
+  peerNet = JSON.parse(peerNetTemp);
+
+  // Distroy default peer before creating new one
+  peer.disconnect();
+  peer.destroy();
+
 }
 
 function sendToPeer(evt, data) {
@@ -169,86 +238,72 @@ function sendToPeer(evt, data) {
   });
 }
 
-function startKey() {
-  console.log('starting key');
+function updateDimFields(evt) {
+  var element = evt.srcElement;
+  var elementId = element.id;
+  var size = element.value;
+  var targetElement = null;
+  
+  evt.preventDefault();
 
-  resetCanvas('color')
-  setImageData();
+  switch (elementId) {
+    case 'colorwidth':
+      targetElement = document.getElementById('colorheight');
+      targetElement.value = (1080 * size) / 1920;
+    break;
 
-  if(kinect.open()) {
-      kinect.on('multiSourceFrame', function(frame) {
+    case 'colorheight': 
+      targetElement = document.getElementById('colorwidth');
+      targetElement.value = (1920 * size) / 1080;
+    break;
 
-        if(busy) {
-          return;
-        }
-        busy = true;
+    case 'depthwidth':
+      targetElement = document.getElementById('depthheight');
+      targetElement.value = (424 * size) / 512;
+    break;
 
-        var closestBodyIndex = getClosestBodyIndex(frame.body.bodies);
-        if(closestBodyIndex !== trackedBodyIndex) {
-          if(closestBodyIndex > -1) {
-            kinect.trackPixelsForBodyIndices([closestBodyIndex]);
-          } else {
-            kinect.trackPixelsForBodyIndices(false);
-          }
-        }
-        else {
-          if (closestBodyIndex > -1) {
-            if (frame.bodyIndexColor.bodies[closestBodyIndex].buffer) {
-
-              newPixelData = frame.bodyIndexColor.bodies[closestBodyIndex].buffer
-
-              for (var i = 0; i < imageDataSize; i++) {
-                imageDataArray[i] = newPixelData[i];
-              }
-
-              drawAndSendImage('key', 'png');
-            }
-          }
-        }
-        trackedBodyIndex = closestBodyIndex;
-        busy = false;
-
-      }); // kinect.on
-    } // open
-      kinect.openMultiSourceReader({
-        frameTypes: Kinect2.FrameType.bodyIndexColor | Kinect2.FrameType.body
-      });
+    case 'depthheight':
+      targetElement = document.getElementById('depthwidth');
+      targetElement.value = (512 * size) / 424;
+    break;
+  }
 }
 
-function stopKey() {
-  kinect.closeMultiSourceReader();
-  kinect.removeAllListeners();
-  busy = false;
-}
+function setOutputDimensions(evt) {
+  var element = evt.srcElement;
+  var elementId = element.id;
+  
+  evt.preventDefault();
 
-function startTracking() {
-      if (kinect.open()) {
-         //listen for body frames
-          kinect.on('bodyFrame', function(bodyFrame){
-            sendToPeer('bodyFrame',bodyFrame);
-              // for (var i = 0;  i < bodyFrame.bodies.length; i++) {
-              //      if (bodyFrame.bodies[i].tracked) {
-              //        console.log("Tracked");
-              //        sendToPeer('bodyFrame',bodyFrame);
-              //        console.log(bodyFrame);
-              //      }
-              // }
-          });
+  switch (elementId) {
+    case 'colorsubmit':
+      outputColorW = document.getElementById('colorwidth').value;
+      outputColorH = document.getElementById('colorheight').value;
 
-          //request body frames
-          kinect.openBodyReader();
+      if (canvasState == 'color' || canvasState === null) {
+        resetCanvas('color');
       }
+      console.log(outputColorW, outputColorH);
+    break;
+
+    case 'depthsubmit':
+      outputDepthW = document.getElementById('depthwidth').value;
+      outputDepthH = document.getElementById('depthheight').value;
+      if (canvasState == 'depth' || canvasState === null) {
+        resetCanvas('depth');
+      }
+        console.log(outputDepthW, outputDepthH);
+    break;
+  }
 }
 
-function stopTracking() {
-  kinect.close();
-  kinect.removeAllListeners();
-  busy = false;
-}
+////////////////////////////////////////////////////////////////////////
+//////////////////////////// Kinect2 Feeds ////////////////////////////
 
 function startRGB() {
 
   resetCanvas('color');
+  canvasState = 'color';
   setImageData();
 
   if(kinect.open()) {
@@ -275,6 +330,7 @@ function startRGB() {
 function stopRGB() {
   kinect.closeColorReader();
   kinect.removeAllListeners();
+  canvasState = null;
   busy = false;
 }
 
@@ -282,34 +338,28 @@ function startDepth() {
   console.log("start Depth Camera");
 
   resetCanvas('depth');
+  canvasState = 'depth';
   setImageData();
 
   if(kinect.open()) {
     kinect.on('depthFrame', function(newPixelData){
-            if(busy) {
-              return;
-            }
-            busy = true;
+      if(busy) {
+        return;
+      }
+      busy = true;
 
-            newPixelDataIndex = 0;
-            for (var i = 0; i < imageDataSize; i+=4) {
-              imageDataArray[i] = newPixelData[newPixelDataIndex];
-              imageDataArray[i+1] = newPixelData[newPixelDataIndex];
-              imageDataArray[i+2] = newPixelData[newPixelDataIndex];
-              imageDataArray[i+3] = newPixelData[newPixelDataIndex];
-              newPixelDataIndex++;
-            }
-
-            drawAndSendImage('depth', 'png');
-            busy = false;
-          });
-        }
+      bufferToImage(newPixelData);
+      drawAndSendImage('depth', 'png');
+      busy = false;
+    });
+  }
   kinect.openDepthReader();
 }
 
 function stopDepth() {
   kinect.closeDepthReader();
   kinect.removeAllListeners();
+  canvasState = null;
   busy = false;
 }
 
@@ -317,29 +367,20 @@ function startInfrared() {
   console.log('starting Infrared Camera');
 
   resetCanvas('depth');
+  canvasState = 'depth';
   setImageData();
      
   if(kinect.open()) {
-    kinect.on('infraredFrame', function(imageBuffer){
+    kinect.on('infraredFrame', function(newPixelData){
       
       if(busy) {
         return;
       }
       busy = true;
-
-      var pixelArray = imageData.data;
-      var newPixelData = new Uint8Array(imageBuffer);
-      var depthPixelIndex = 0;
-
-      for (var i = 0; i < imageDataSize; i+=4) {
-        pixelArray[i] = newPixelData[depthPixelIndex];
-        pixelArray[i+1] = newPixelData[depthPixelIndex];
-        pixelArray[i+2] = newPixelData[depthPixelIndex];
-        pixelArray[i+3] = 0xff;
-        depthPixelIndex++;
-      }
-
+      
+      bufferToImage(newPixelData);
       drawAndSendImage('infrared', 'jpeg');
+      
       busy = false;
     });
   }
@@ -352,6 +393,7 @@ function stopInfrared() {
   console.log('stopping Infrared Camera');
   kinect.closeInfraredReader();
   kinect.removeAllListeners();  
+  canvasState = null;
   busy = false;
 }
 
@@ -359,27 +401,18 @@ function startLEInfrared() {
   console.log('starting LE Infrared');
 
   resetCanvas('depth');
+  canvasState = 'depth';
   setImageData();
 
 
   if(kinect.open()) {
-    kinect.on('longExposureInfraredFrame', function(imageBuffer){
+    kinect.on('longExposureInfraredFrame', function(newPixelData){
       if(busy) {
         return;
       }
       busy = true;
       
-      var pixelArray = imageData.data;
-      var newPixelData = new Uint8Array(imageBuffer);
-      var depthPixelIndex = 0;
-      for (var i = 0; i < imageDataSize; i+=4) {
-        pixelArray[i] = newPixelData[depthPixelIndex];
-        pixelArray[i+1] = newPixelData[depthPixelIndex];
-        pixelArray[i+2] = newPixelData[depthPixelIndex];
-        pixelArray[i+3] = 0xff;
-        depthPixelIndex++;
-      }
-      
+      bufferToImage(newPixelData);
       drawAndSendImage('LEinfrared', 'jpeg');
 
       busy = false;
@@ -394,12 +427,158 @@ function stopLEInfrared() {
   console.log('stopping LE Infrared');
   kinect.closeLongExposureInfraredReader();
   kinect.removeAllListeners();
+  canvasState = null;
   busy = false;
 }
+
+function startSkeletonTracking() {
+  console.log('starting skeleton');
+  console.log('send all', sendAllBodies);
+  
+  resetCanvas('depth');
+  canvasState = 'depth';
+
+  if(kinect.open()) {
+    kinect.on('bodyFrame', function(bodyFrame){
+      if(sendAllBodies) {
+        console.log('sendingall');
+        sendToPeer('bodyFrame', bodyFrame);
+      }
+
+      context.clearRect(0, 0, canvas.width, canvas.height);
+      outputContext.clearRect(0, 0, outputCanvas.width, outputCanvas.height);
+      var index = 0;
+      bodyFrame.bodies.forEach(function(body){
+        if(body.tracked) {
+          console.log('tracked');
+          if (!sendAllBodies) {
+            console.log('tracked only');
+            sendToPeer('trackedBodyFrame', body);
+          }
+          for(var jointType in body.joints) {
+            var joint = body.joints[jointType];
+            context.fillStyle = colors[index];
+            context.fillRect(joint.depthX * canvas.width, joint.depthY * canvas.height, 10, 10);
+            outputContext.fillStyle = colors[index];
+            outputContext.fillRect(joint.depthX * outputCanvas.width, joint.depthY * outputCanvas.height, 10, 10);
+          }
+          //draw hand states
+          updateHandState(context, body.leftHandState, body.joints[Kinect2.JointType.handLeft]);
+          updateHandState(outputContext, body.leftHandState, body.joints[Kinect2.JointType.handLeft]);
+          updateHandState(context, body.rightHandState, body.joints[Kinect2.JointType.handRight]);
+          updateHandState(outputContext, body.rightHandState, body.joints[Kinect2.JointType.handRight]);
+
+          index++;
+
+        }
+      });
+    });
+    kinect.openBodyReader();
+      }
+
+}
+
+function stopSkeletonTracking() {
+  console.log('stopping skeleton');
+  kinect.closeBodyReader();
+  kinect.removeAllListeners();
+  canvasState = null;
+
+}
+
+
+function startKey() {
+  console.log('starting key');
+
+  resetCanvas('color');
+  canvasState = 'color';
+  setImageData();
+
+  if(kinect.open()) {
+      kinect.on('multiSourceFrame', function(frame) {
+
+        if(busy) {
+          return;
+        }
+        busy = true;
+
+        var closestBodyIndex = getClosestBodyIndex(frame.body.bodies);
+        if(closestBodyIndex !== trackedBodyIndex) {
+          if(closestBodyIndex > -1) {
+            kinect.trackPixelsForBodyIndices([closestBodyIndex]);
+          } else {
+            kinect.trackPixelsForBodyIndices(false);
+          }
+        }
+        else {
+          if (closestBodyIndex > -1) {
+            if (frame.bodyIndexColor.bodies[closestBodyIndex].buffer) {
+
+              newPixelData = frame.bodyIndexColor.bodies[closestBodyIndex].buffer;
+
+              for (var i = 0; i < imageDataSize; i++) {
+                imageDataArray[i] = newPixelData[i];
+              }
+
+              drawAndSendImage('key', 'png');
+            }
+          }
+        }
+        trackedBodyIndex = closestBodyIndex;
+        busy = false;
+
+      }); // kinect.on
+    } // open
+      kinect.openMultiSourceReader({
+        frameTypes: Kinect2.FrameType.bodyIndexColor | Kinect2.FrameType.body
+      });
+}
+
+function stopKey() {
+  kinect.closeMultiSourceReader();
+  kinect.removeAllListeners();
+  canvasState = null;
+  busy = false;
+}
+
+function startTracking() {
+      if (kinect.open()) {
+         //listen for body frames
+          kinect.on('bodyFrame', function(bodyFrame){
+            sendToPeer('bodyFrame',bodyFrame);
+              // for (var i = 0;  i < bodyFrame.bodies.length; i++) {
+              //      if (bodyFrame.bodies[i].tracked) {
+              //        console.log("Tracked");
+              //        sendToPeer('bodyFrame',bodyFrame);
+              //        console.log(bodyFrame);
+              //      }
+              // }
+          });
+
+          //request body frames
+          kinect.openBodyReader();
+      }
+}
+
+function stopTracking() {
+  kinect.close();
+  kinect.removeAllListeners();
+  canvasState = null;
+  busy = false;
+}
+
+
+
+
+
+
+
+
 
 function startFHJoint() {
 
   resetCanvas('color');
+  canvasState = 'color';
   setImageData();
   
   trackedBodyIndex = -1;
@@ -418,7 +597,7 @@ function startFHJoint() {
         imageDataArray[i] = newPixelData[i];
       }
 
-      drawAndSendImage('fhcolor', 'jpeg');
+      //drawAndSendImage('fhcolor', 'jpeg');
   
       // get closest body
       var closestBodyIndex = getClosestBodyIndex(frame.body.bodies);
@@ -470,6 +649,7 @@ function stopFHJoint() {
   console.log('stopping FHJoint');
   kinect.closeMultiSourceReader();
   kinect.removeAllListeners();
+  canvasState = null;
   busy = false;
 }
 
@@ -477,6 +657,7 @@ function startScaleUser() {
   console.log('start scale user');
 
   resetCanvas('color');
+  canvasState = 'color';
   setImageData();
 
   trackedBodyIndex = -1;
@@ -564,52 +745,11 @@ function stopScaleUser() {
   console.log('stop scale user');
   kinect.closeMultiSourceReader();
   kinect.removeAllListeners();
+  canvasState = null;
   busy = false;
 }      
 
-function startSkeletonTracking() {
-  console.log('starting skeleton');
-  
-  resetCanvas('depth');
 
-  if(kinect.open()) {
-    kinect.on('bodyFrame', function(bodyFrame){
-
-      context.clearRect(0, 0, canvas.width, canvas.height);
-      outputContext.clearRect(0, 0, canvas.width, canvas.height);
-      var index = 0;
-      bodyFrame.bodies.forEach(function(body){
-        if(body.tracked) {
-          sendToPeer('bodyFrame', body);
-          for(var jointType in body.joints) {
-            var joint = body.joints[jointType];
-            context.fillStyle = colors[index];
-            context.fillRect(joint.depthX * canvas.width, joint.depthY * canvas.height, 10, 10);
-            outputContext.fillStyle = colors[index];
-            outputContext.fillRect(joint.depthX * outputCanvas.width, joint.depthY * outputCanvas.height, 10, 10);
-          }
-          //draw hand states
-          updateHandState(context, body.leftHandState, body.joints[Kinect2.JointType.handLeft]);
-          updateHandState(outputContext, body.leftHandState, body.joints[Kinect2.JointType.handLeft]);
-          updateHandState(context, body.rightHandState, body.joints[Kinect2.JointType.handRight]);
-          updateHandState(outputContext, body.rightHandState, body.joints[Kinect2.JointType.handRight]);
-
-          index++;
-
-        }
-      });
-    });
-    kinect.openBodyReader();
-      }
-
-}
-
-function stopSkeletonTracking() {
-  console.log('stopping skeleton');
-  kinect.closeBodyReader();
-  kinect.removeAllListeners();
-
-}
 
 function loadFile(e) {
   window.location.href = e.target.files[0].path;
@@ -624,20 +764,17 @@ function setImageData() {
 function resetCanvas(size) {
   context.clearRect(0, 0, canvas.width, canvas.height);
   outputContext.clearRect(0, 0, outputCanvas.width, outputCanvas.height);
-  //sendToPeer('clearCanvas', {});
   
   if (size == 'depth') {
     canvas.width = DEPTHWIDTH;
     canvas.height = DEPTHHEIGHT;
-    outputCanvas.width = DEPTHWIDTH;
-    outputCanvas.height = DEPTHHEIGHT;
-    sendToPeer('framesize', {'size': 'depth'});
+    outputCanvas.width = outputDepthW;
+    outputCanvas.height = outputDepthH;
   } else if (size == 'color') {
     canvas.width = COLORWIDTH;
     canvas.height = COLORHEIGHT; 
-    outputCanvas.width = COLORWIDTH;
-    outputCanvas.height = COLORHEIGHT;
-    sendToPeer('framesize', {'size': 'color'});
+    outputCanvas.width = outputColorW;
+    outputCanvas.height = outputColorH;
   }
 }
     
@@ -652,6 +789,20 @@ function drawAndSendImage(frameType, imageType) {
   dataToSend = {'name': frameType, 'imagedata': outputCanvasData};
 
   sendToPeer('frame', dataToSend);
+}
+
+
+// process depth and infrared cameras buffer
+function bufferToImage(newPixelData){
+  var j = 0;
+
+  for (var i = 0; i < imageDataSize; i+=4) {
+    imageDataArray[i] = newPixelData[j];
+    imageDataArray[i+1] = newPixelData[j];
+    imageDataArray[i+2] = newPixelData[j];
+    imageDataArray[i+3] = 0xff; // set alpha channel at full opacity
+    j++;
+  }
 }
 
 function getClosestBodyIndex(bodies) {
@@ -722,4 +873,61 @@ function drawHand(context, jointPoint, handColor) {
   context.fill();
   context.closePath();
   context.globalAlpha = 1;
+}
+
+
+// Get IP Address. Taken from http://net.ipcalf.com/
+if (RTCPeerConnection) (function () {
+    var rtc = new RTCPeerConnection({iceServers:[]});
+    if (1 || window.mozRTCPeerConnection) {      // FF [and now Chrome!] needs a channel/stream to proceed
+        rtc.createDataChannel('', {reliable:false});
+    }
+    
+    rtc.onicecandidate = function (evt) {
+        // convert the candidate to SDP so we can run it through our general parser
+        // see https://twitter.com/lancestout/status/525796175425720320 for details
+        if (evt.candidate) grepSDP("a="+evt.candidate.candidate);
+    };
+    rtc.createOffer(function (offerDesc) {
+        grepSDP(offerDesc.sdp);
+        rtc.setLocalDescription(offerDesc);
+    }, function (e) { console.warn("offer failed", e); });
+    
+    
+    var addrs = Object.create(null);
+    addrs["0.0.0.0"] = false;
+    function updateDisplay(newAddr) {
+        if (newAddr in addrs) return;
+        else addrs[newAddr] = true;
+        var ips = Object.getOwnPropertyNames(addrs).sort();
+
+        // find local ip address
+        for (i = 0; i < ips.length; i++) {
+          if (ips[i].includes('192')) {
+            var displayAddrs = ips[i];
+          }
+          
+        }
+        //var displayAddrs = Object.keys(addrs).filter(function (k) { return addrs[k]; });
+        document.getElementById('ipaddress').textContent = displayAddrs;
+    }
+    
+    function grepSDP(sdp) {
+        var hosts = [];
+        sdp.split('\r\n').forEach(function (line) { // c.f. http://tools.ietf.org/html/rfc4566#page-39
+            if (~line.indexOf("a=candidate")) {     // http://tools.ietf.org/html/rfc4566#section-5.13
+                var parts = line.split(' '),        // http://tools.ietf.org/html/rfc5245#section-15.1
+                    addr = parts[4],
+                    type = parts[7];
+                if (type === 'host') updateDisplay(addr);
+            } else if (~line.indexOf("c=")) {       // http://tools.ietf.org/html/rfc4566#section-5.7
+                var parts = line.split(' '),
+                    addr = parts[2];
+                updateDisplay(addr);
+            }
+        });
+    }
+})(); else {
+    document.getElementById('ipaddress').innerHTML = "<code>ifconfig | grep inet | grep -v inet6 | cut -d\" \" -f2 | tail -n1</code>";
+    document.getElementById('ipaddress').nextSibling.textContent = "In Chrome and Firefox your IP should display automatically, by the power of WebRTCskull.";
 }
