+++ src/p5.kinect2.js
@@ -76,9 +76,13 @@ p5.Kinect2 = function(peerid, network) {
           this.img.elt.src = data.imagedata;
           this.callback(this.img);
         break;
+
+        case 'bodyFrame':
+          this.callback(data);
+        break;
  
         // If skeleton data, draw skeleton
-        case 'bodyFrame':
+        case 'trackedBodyFrame':
           this.body = data;
           this.callback(data);
         break;
@@ -111,7 +115,12 @@ p5.Kinect2 = function(peerid, network) {
     this._setFeed('le-infrared');
   };
 
-  this.startSkeleton = function(callback) {
+  this.startBodies = function(callback) {
+    this.callback = callback;
+    this._setFeed('body');
+  };
+
+  this.startTrackedBodies = function(callback) {
     this.callback = callback;
     this._setFeed('skeleton');
   };
@@ -121,10 +130,10 @@ p5.Kinect2 = function(peerid, network) {
     this._setFeed('key');
   };
 
-  this.startScale = function(callback) {
-    this.callback = callback;
-    this._setFeed('scale');
-  };
+  // this.startScale = function(callback) {
+  //   this.callback = callback;
+  //   this._setFeed('scale');
+  // };
 
   this.startFloorHeight = function(callback) {
     this.callback = callback;
